// lib/contentful.d.ts
export interface I3ZPkEj1KqeSn4QdsdnNko3Fields {
  /** Personalized Hero Name */
  name: string;

  /** Hero Options */
  unfrmOptP13nList?: IHero[] | undefined;
}

export interface I3ZPkEj1KqeSn4QdsdnNko3 extends Entry<I3ZPkEj1KqeSn4QdsdnNko3Fields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: '3zPkEj1KqeSn4QdsdnNKO3';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface ICtaFields {
  /** Heading */
  heading?: string | undefined;

  /** Subheading */
  subheading?: string | undefined;

  /** Button Link */
  buttonLink?: string | undefined;

  /** Button Text */
  buttonText?: string | undefined;

  /** Button Image */
  buttonImage?: Asset | undefined;
}

export interface ICta extends Entry<ICtaFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'cta';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IHeroFields {
  /** Title */
  title: string;

  /** Description */
  description: string;

  /** Button Text */
  buttonText?: string | undefined;

  /** image */
  image?: Asset | undefined;

  /** Intent Tags */
  unfrmOptIntentTag?: Record<string, any> | undefined;

  /** Button Link Slug */
  buttonLinkSlug?: string | undefined;
}

export interface IHero extends Entry<IHeroFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'hero';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IPageFields {
  /** Title */
  title?: string | undefined;

  /** Slug */
  slug?: string | undefined;

  /** Components */
  components?: (ICta | IHero | I3ZPkEj1KqeSn4QdsdnNko3 | IRegistrationForm | ITalksList)[] | undefined;
}

export interface IPage extends Entry<IPageFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'page';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IRegistrationFormFields {
  /** Heading */
  heading?: string | undefined;

  /** ButtonText */
  buttonText?: string | undefined;

  /** Registered Text */
  registeredText?: string | undefined;
}

export interface IRegistrationForm extends Entry<IRegistrationFormFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'registrationForm';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface ITalkFields {
  /** Title */
  title?: string | undefined;

  /** Description */
  description?: Document | undefined;

  /** Intent Tags */
  unfrmOptIntentTag?: Record<string, any> | undefined;
}

export interface ITalk extends Entry<ITalkFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'talk';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface ITalksListFields {
  /** Title */
  title?: string | undefined;

  /** Title When Personalized */
  p13nTitle?: string | undefined;

  /** NumberToShow */
  count?: number | undefined;

  /** Register Button Text */
  registerButtonText?: string | undefined;
}

export interface ITalksList extends Entry<ITalksListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'talksList';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}